Refactor sprite handling into manager
